requirements.txt
torch
torchvision
opencv-python
numpy
pillow
scikit-image
matplotlib
streamlit
tqdm